STIV pipeline chores: gitignore updates, float64 cast fix, Station Number moved to Processing, drop LSPIV fusion
Moves the Station Number control from the Results form to Processing
(where it belongs alongside the other pre-analysis settings), drops the
now-unused LSPIV fusion path from the STIV pipeline in favor of a
standalone angle profile, casts cached stats arrays to float64, and adds
a few gitignore entries (venv symlink, .superpowers/ scratch dir).

Squashed from 6 commits (2026-07-17).

+++ river/core/compute_section.py
@@ -1506,9 +1506,12 @@ def update_current_x_section(
         existing_cache = x_sections[current_x_section].get("_stats_cache")
 
         if existing_cache and existing_cache.get("geometry_hash") == geometry_hash:
-            streamwise_vel_frames = np.array(existing_cache["streamwise_vel_frames"])
-            gradient_frames = np.array(existing_cache["gradient_frames"])
-            dense_distances = np.array(existing_cache["dense_distances"])
+            # dtype=float64 is required: NaN values in the cache round-trip through the
+            # Electron layer as JSON null (JSON has no NaN), which np.array() would
+            # otherwise load as dtype=object instead of coercing back to nan.
+            streamwise_vel_frames = np.array(existing_cache["streamwise_vel_frames"], dtype=np.float64)
+            gradient_frames = np.array(existing_cache["gradient_frames"], dtype=np.float64)
+            dense_distances = np.array(existing_cache["dense_distances"], dtype=np.float64)
         else:
             streamwise_vel_frames, gradient_frames = _compute_stats_cache(
                 piv_results,
